Seperated Bounding box from game_entity.cpp
Moved the enitreity of game_entity.cpp to bounding_box.cpp
as all code was concerned with the bounding box and game entities are
just an interface.

Bounding boxes will also be important when implementing the quad (or R)
tree however game enitty should not be a depency for the tree.